requirements.txt
requests==2.31.0
bs4==0.0.2
langdetect==1.0.9
pandas==2.0.3
apache-airflow-providers-amazon==8.17.0